[change] シリアライザーのバージョニング
