+++ backend/app/app/controllers/v1/users_controller.rb
@@ -3,7 +3,7 @@ module V1
     def show
       user = User.find_by!(name: params[:name])
       options = {}
-      json_string = UserSerializer.new(user, options).serialized_json
+      json_string = V1::UserSerializer.new(user, options).serialized_json
       render json: json_string, status: :ok
     end
   end
