--- backend/app/app/serializers/post_serializer.rb
@@ -1,7 +1,0 @@
-class PostSerializer
-  include FastJsonapi::ObjectSerializer
-
-  belongs_to :user
-
-  attributes :title, :description, :created_at, :updated_at
-end
