--- backend/app/app/serializers/question_serializer.rb
@@ -1,7 +1,0 @@
-class QuestionSerializer
-  include FastJsonapi::ObjectSerializer
-
-  belongs_to :post
-
-  attributes :statement, :status
-end
