--- backend/app/app/serializers/user_post_serializer.rb
@@ -1,5 +1,0 @@
-class UserPostSerializer
-  include FastJsonapi::ObjectSerializer
-
-  attributes :title, :description, :status, :created_at, :updated_at
-end
