--- backend/app/app/serializers/user_serializer.rb
@@ -1,5 +1,0 @@
-class UserSerializer
-  include FastJsonapi::ObjectSerializer
-
-  attributes :name, :nickname, :profile
-end
