+++ backend/app/app/serializers/v1/post_serializer.rb
@@ -1,0 +1,9 @@
+module V1
+  class PostSerializer
+    include FastJsonapi::ObjectSerializer
+
+    belongs_to :user
+
+    attributes :title, :description, :created_at, :updated_at
+  end
+end
