+++ backend/app/app/serializers/v1/question_serializer.rb
@@ -1,0 +1,9 @@
+module V1
+  class QuestionSerializer
+    include FastJsonapi::ObjectSerializer
+
+    belongs_to :post
+
+    attributes :statement, :status
+  end
+end
