+++ backend/app/app/serializers/v1/user_post_serializer.rb
@@ -1,0 +1,7 @@
+module V1
+  class UserPostSerializer
+    include FastJsonapi::ObjectSerializer
+
+    attributes :title, :description, :status, :created_at, :updated_at
+  end
+end
