+++ backend/app/app/serializers/v1/user_serializer.rb
@@ -1,0 +1,7 @@
+module V1
+  class UserSerializer
+    include FastJsonapi::ObjectSerializer
+
+    attributes :name, :nickname, :profile
+  end
+end
